+++ src/material.rs
@@ -11,7 +11,7 @@ pub fn random_point_in_sphere() -> Vec3 {
     loop {
         // select random points inside the unit cube until a selected point
         // is contained inside the unit sphere
-        let p = Vec3::new(rng.gen::<f64>(), rng.gen::<f64>(), rng.gen::<f64>()) * 2.0 - unit;
+        let p = Vec3::new(rng.gen::<f32>(), rng.gen::<f32>(), rng.gen::<f32>()) * 2.0 - unit;
         if p.squared_length() < 1.0 {
             return p;
         }
@@ -23,7 +23,7 @@ pub fn reflect(v: Vec3, n: Vec3) -> Vec3 {
     v - n * v.dot(n) * 2.0
 }
 
-pub fn refract(v: Vec3, n: Vec3, ni_over_nt: f64) -> Option<Vec3> {
+pub fn refract(v: Vec3, n: Vec3, ni_over_nt: f32) -> Option<Vec3> {
     // refract ray direction `v` across normal `n`
     // `ni_over_nt` is the ratio of material refractive indices
     let uv = v.unit();
@@ -37,7 +37,7 @@ pub fn refract(v: Vec3, n: Vec3, ni_over_nt: f64) -> Option<Vec3> {
     }
 }
 
-pub fn schlick(cosine: f64, refractive_idx: f64) -> f64 {
+pub fn schlick(cosine: f32, refractive_idx: f32) -> f32 {
     // schlick's approximation for the fresnel equations
     let r0 = ((1.0 - refractive_idx) / (1.0 + refractive_idx)).powi(2);
     r0 + (1.0 - r0) * (1.0 - cosine).powi(5)
@@ -70,12 +70,12 @@ impl Material for Lambertian {
 
 #[derive(Debug, Copy, Clone)]
 pub struct Metal {
-    pub fuzz: f64,
+    pub fuzz: f32,
     pub albedo: Vec3,
 }
 
 impl Metal {
-    pub fn new(fuzz: f64, albedo: Vec3) -> Metal {
+    pub fn new(fuzz: f32, albedo: Vec3) -> Metal {
         Metal {
             fuzz: if fuzz < 1.0 { fuzz } else { 1.0 },
             albedo,
@@ -98,13 +98,13 @@ impl Material for Metal {
 
 #[derive(Debug, Copy, Clone)]
 pub struct Dielectric {
-    pub ri: f64,
-    pub fuzz: f64,
+    pub ri: f32,
+    pub fuzz: f32,
     pub albedo: Vec3,
 }
 
 impl Dielectric {
-    pub fn new(ri: f64, fuzz: f64, albedo: Vec3) -> Dielectric {
+    pub fn new(ri: f32, fuzz: f32, albedo: Vec3) -> Dielectric {
         Dielectric { ri, fuzz, albedo }
     }
 }
@@ -133,7 +133,7 @@ impl Material for Dielectric {
         // or refracted light based on `reflect_prob`
         if let Some(refracted) = refract(r.direction, outward_normal, ni_over_nt) {
             let reflect_prob = schlick(cos, self.ri);
-            if rand::thread_rng().gen::<f64>() >= reflect_prob {
+            if rand::thread_rng().gen::<f32>() >= reflect_prob {
                 let scattered =
                     Ray::new(record.p, refracted + random_point_in_sphere() * self.fuzz);
                 return Some((scattered, attenuation));
